fix error class
