+++ packages/typescript/errors.ts
@@ -1,5 +1,35 @@
-export class AccessDenied extends Error {}
+import { Response } from 'node-fetch';
+export class AccessDenied extends Error {
+    constructor(message: string) {
+        super(message);
 
-export class ServerSideError extends Error {}
+        Object.setPrototypeOf(this, AccessDenied.prototype);
+    }
+}
 
-export class KeyError extends Error {}
+export class ServerSideError extends Error {
+    constructor(message: string) {
+        super(message);
+
+        Object.setPrototypeOf(this, ServerSideError.prototype);
+    }
+}
+
+export class KeyError extends Error {
+    constructor(message: string) {
+        super(message);
+
+        Object.setPrototypeOf(this, KeyError.prototype);
+    }
+}
+
+export class HTTPResponseError extends Error {
+    response: Response;
+    constructor(response: Response) {
+        super(
+            `HTTP Error Response: ${response.status} ${response.statusText}`
+        );
+        this.response = response;
+        Object.setPrototypeOf(this, HTTPResponseError.prototype);
+    }
+}
